feat: Add speaker inference and UI_json inference
Add speaker inference and UI_json

+++ Utils/wer_eval.py
@@ -17,7 +17,7 @@ def eval_asr(asr_truth,asr_predict_result):
        wer = Lev.distance(asr_predict_result[i], asr_truth[i]) / float(len(asr_truth[i]))
 #       print(f"{wer}")
        wer_list.append(wer)
-    ASR_wer_avg = np.sum(wer_list)/len(wer_list)        
+    ASR_wer_avg = float(np.sum(wer_list)/len(wer_list))        
     return ASR_wer_avg, wer_list
